- В статистику было добавлено кол-во предупреждений - Теперь все меню работают по следующему принципу (Сначала выделяешь на ПКМ, а потом нажимаешь еще раз, он уже выполняет действие). По типу дабл клика в Windows (Когда ты по папкам переходишь) - В настройках интерфейса была добавлена возможность вкл/выкл курсора, при пользовании обычных меню по умолчанию (Например, если вы отключите курсор , то это вам даст возможность выбирать клавиатурой необходимые айтемы и при этом крутить камерой) - В настройках интерфейса добавлена возможность отключить уведомления над картой - Убран дополнительный налог на ЛСК - Фикс бага, с тем, когда ты печатаешь и открывается меню с жалобой - Фикс бага, с пропаданием чата - Фикс розыгрыша машин (Завтра с 7.30 по Мск, в рандомное время будет разыгран 1 дополнительный автомобиль) - Фиксы различных багов и опечаток

+++ packages/dnet/modules/events.js
@@ -2106,6 +2106,17 @@ mp.events.addRemoteCounted('server:invader:delAd', (player, id) => {
 
     user.addPayDayMoney(player, 100, 'Отредактировал объявление');
 
+    mysql.executeQuery(`DELETE FROM rp_inv_ad WHERE id = ${methods.parseInt(id)}`);
+    player.notify('~b~Вы удалили объявление #' + id);
+    player.notify('~g~Вы получили премию в $100 за отредактированное объявление');
+});
+
+mp.events.addRemoteCounted('server:invader:delAdT', (player, id) => {
+    if (!user.isLogin(player))
+        return;
+
+    user.addPayDayMoney(player, 100, 'Отредактировал объявление');
+
     mysql.executeQuery(`DELETE FROM rp_inv_ad_temp WHERE id = ${methods.parseInt(id)}`);
     player.notify('~b~Вы удалили объявление #' + id);
     player.notify('~g~Вы получили премию в $100 за отредактированное объявление');
